+++ src/main/java/com/myce/expo/controller/ExpoController.java
@@ -16,21 +16,14 @@ import java.util.List;
 @RequestMapping("/api/expos")
 @RequiredArgsConstructor
 public class ExpoController {
-  private final ExpoService exposervice;
+    private final ExpoService exposervice;
 
-  // 관리하고 있는 박람회 리스트
-  @GetMapping("/my")
-  public ResponseEntity<List<Long>> getMyExpos(@AuthenticationPrincipal CustomUserDetails customUserDetails){
-    Long memberId = customUserDetails.getMemberId();
-    return ResponseEntity.ok(exposervice.getMyExpos(memberId));
-  }
-
-  // 박람회 등록
-  @PostMapping
-  public ResponseEntity<Long> saveExpo(@AuthenticationPrincipal CustomUserDetails customUserDetails,
-      @RequestBody @Valid ExpoRegistrationRequest expoRegistrationRequest){
-    Long memberId = customUserDetails.getMemberId();
-    exposervice.saveExpo(memberId, expoRegistrationRequest);
-    return ResponseEntity.status(HttpStatus.CREATED).build();
-  }
+    // 박람회 등록
+    @PostMapping
+    public ResponseEntity<Long> saveExpo(@AuthenticationPrincipal CustomUserDetails customUserDetails,
+                                         @RequestBody @Valid ExpoRegistrationRequest expoRegistrationRequest) {
+        Long memberId = customUserDetails.getMemberId();
+        exposervice.saveExpo(memberId, expoRegistrationRequest);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
+    }
 }
