+++ src/main/java/com/myce/expo/service/ExpoService.java
@@ -5,7 +5,6 @@ import com.myce.expo.dto.ExpoRegistrationRequest;
 import java.util.List;
 
 public interface ExpoService {
-  void saveExpo(Long memberId, ExpoRegistrationRequest request);
-  List<Long> getMyExpos(Long memberId);
+    void saveExpo(Long memberId, ExpoRegistrationRequest request);
 }
