+++ src/main/java/com/myce/expo/service/impl/ExpoServiceImpl.java
@@ -27,49 +27,43 @@ import java.util.List;
 @Transactional
 @RequiredArgsConstructor
 public class ExpoServiceImpl implements ExpoService {
-  private final MemberRepository memberRepository;
-  private final ExpoRepository expoRepository;
-  private final CategoryRepository  categoryRepository;
-  private final BusinessProfileRepository businessProfileRepository;
+    private final MemberRepository memberRepository;
+    private final ExpoRepository expoRepository;
+    private final CategoryRepository categoryRepository;
+    private final BusinessProfileRepository businessProfileRepository;
 
-  @Override
-  public void saveExpo(Long memberId, ExpoRegistrationRequest request) {
-    // 로그인한 사용자
-    Member member = memberRepository.findById(memberId)
-        .orElseThrow(() -> new CustomException(CustomErrorCode.MEMBER_NOT_EXIST));
+    @Override
+    public void saveExpo(Long memberId, ExpoRegistrationRequest request) {
+        // 로그인한 사용자
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new CustomException(CustomErrorCode.MEMBER_NOT_EXIST));
 
-    // expo 객체 생성
-    Expo expo = ExpoMapper.toEntity(request, member);
+        // expo 객체 생성
+        Expo expo = ExpoMapper.toEntity(request, member);
 
-    // 카테고리 추가
-    for (Long categoryId : request.getCategoryIds()) {
-      Category category = categoryRepository.findById(categoryId)
-          .orElseThrow(() -> new CustomException(CustomErrorCode.CATEGORY_NOT_EXIST));
+        // 카테고리 추가
+        for (Long categoryId : request.getCategoryIds()) {
+            Category category = categoryRepository.findById(categoryId)
+                    .orElseThrow(() -> new CustomException(CustomErrorCode.CATEGORY_NOT_EXIST));
 
-      ExpoCategory expoCategory = ExpoCategory.builder()
-          .category(category)
-          .expo(expo)
-          .build();
+            ExpoCategory expoCategory = ExpoCategory.builder()
+                    .category(category)
+                    .expo(expo)
+                    .build();
 
-      expo.getExpoCategories().add(expoCategory);
+            expo.getExpoCategories().add(expoCategory);
+        }
+
+        // 박람회 등록(저장)
+        Expo savedExpo = expoRepository.save(expo);
+
+        // 등록 신청한 회사 정보 저장
+        ExpoRegistrationCompanyRequest company = request.getExpoRegistrationCompanyRequest();
+
+        BusinessProfile businessProfile = BusinessProfileMapper.toEntity(company, savedExpo.getId());
+
+        businessProfileRepository.save(businessProfile);
     }
 
-    // 박람회 등록(저장)
-    Expo savedExpo = expoRepository.save(expo);
 
-    // 등록 신청한 회사 정보 저장
-    ExpoRegistrationCompanyRequest company = request.getExpoRegistrationCompanyRequest();
-
-    BusinessProfile businessProfile = BusinessProfileMapper.toEntity(company, savedExpo.getId());
-
-    businessProfileRepository.save(businessProfile);
-  }
-
-  @Override
-  public List<Long> getMyExpos(Long memberId) {
-    List<Expo> expos = expoRepository.findByMemberIdAndStatusIn(memberId, ExpoStatus.ACTIVE_STATUSES);
-    return expos.stream()
-            .map(Expo::getId)
-            .toList();
-  }
 }
